backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-dotenv==1.0.1
email-validator==2.3.0

# Resume parsing
PyPDF2==3.0.1
python-docx==1.1.2

# AI/ML (optional - for future)
# openai==1.51.2

# Database (future)
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.35
